# Save one utf8 conversion in KeywordFieldMapper.