@mikemccand 2016-08-17
Unfortunately, this change caused a large drop in indexing performance for the geonames documents: https://benchmarks-old.elastic.co/index.html#indexing

What's very strange is, I can easily reproduce the big indexing throughput loss on my 72 core box, before and after this change, but on 8 core Haswell and Skylake boxes, there is no measurable change in indexing performance.  I can't explain that.

The change looks quite innocuous, but when I stare at Lucene's sources for how it handles the `String` vs `BytesRef` cases, there are some interesting differences, e.g. the `BinaryTokenStream` does not have an `OffsetAttribute`, the `BytesTermAttribute.clear` does not reuse the `byte[]` buffer while `CharTermAttributeImpl.clear` does reuse its `char[] termBuffer`.

Maybe we are confusing hotspot by having two hot implementations for `termAtt.getBytesRef` called from `TermsHashPerField.add`?

Maybe we should back this out until we understand what's happening?  I think the motivation for the change is very good, but I can't explain why it's giving such a big drop on a high core count box...