@jpountz 2016-08-17
Thanks for catching the slow down. +1 to back it out. I'll do it tomorrow unless someone is able to do it before.