@mikemccand 2016-08-17
I'm sorry to be the bearer of bad news here: I think it's ridiculous we cannot make a clear improvement in ES for this reason.  I hope we can get to the bottom of it.  It ought to be safe to index binary terms directly in Lucene.

Anyway, @rmuir has been helping me try to isolate the cause for the slowdown here, and at least we have made some progress: if I run the benchmark with `-XX:TieredStopAtLevel=1` (disables Hotspot's C2 compilation "efforts") then:

```
after: Indexer: 8647880 docs: 320.33 sec [26997.1 dps, 8.7 MB/sec]
before: Indexer: 8647880 docs: 343.35 sec [25186.9 dps, 8.1 MB/sec]
```

So, happily, the change ("after") is a bit faster, indexing all geonames docs, if we force only C1 (essentially `-client` as @rmuir summarizes it)...