@rmuir 2016-08-18
Honestly, i don't get it, maybe its a compiler bug, or perhaps related to lock elision or something else on your machine (though, we tried disabling EA).

Because i think even if this part of indexwriter is compiled inefficiently, it shouldn't create such a massive performance regression for the app as a whole, so something truly crazy is going on.