@mikemccand 2016-08-18
I also tested JDK-9 EA: still a big slowdown.

And I upgraded my kernel from 4.6.x to 4.7.0: still a big slowdown.

It's so weird that single socket boxes, at least the two 8 core boxes I've tested on, don't show any slowdown.