@mikemccand 2016-08-18
OK @rmuir and I finally found the smoking gun (using JFR, but running with `-XX:-Inline` so you see all methods).  With this tiny change to Lucene 6.1.0, this change has the same performance as before on my 2 socket (72 core) box:

```
mike@beast2:/l/61/lucene/core$ git diff
diff --git a/lucene/core/src/java/org/apache/lucene/analysis/TokenStream.java b/lucene/core/src/java/org/apache/lucene/analysis/TokenStream.java
index 6a78e1c..1f89f06 100644
--- a/lucene/core/src/java/org/apache/lucene/analysis/TokenStream.java
+++ b/lucene/core/src/java/org/apache/lucene/analysis/TokenStream.java
@@ -177,10 +177,10 @@ public abstract class TokenStream extends AttributeSource implements Closeable {
    */
   public void end() throws IOException {
     clearAttributes(); // LUCENE-3849: don't consume dirty atts
-    PositionIncrementAttribute posIncAtt = getAttribute(PositionIncrementAttribute.class);
-    if (posIncAtt != null) {
-      posIncAtt.setPositionIncrement(0);
-    }
+    //PositionIncrementAttribute posIncAtt = getAttribute(PositionIncrementAttribute.class);
+    //if (posIncAtt != null) {
+    //posIncAtt.setPositionIncrement(0);
+    //}
   }
 
   /**
```

But this is not committable to Lucene ... so we need to figure out how to properly fix it...