@rmuir 2016-08-18
https://issues.apache.org/jira/browse/LUCENE-7419