@jpountz 2016-08-25
I just pushed this change back in.