@jpountz 2016-08-25
Thank you @mikemccand and @rmuir for looking into it!